create helper utils, addcomponent, updatecomponent, and deletecomponent
