+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "reactype",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "description": "Prototyping tool for React/Typescript Applications.",
   "main": "main.js",
   "contributors": [
