+++ src/types/actionTypes.ts
@@ -1,10 +1,16 @@
 // ** Application specific action types
 export const LOAD_INIT_DATA: string = 'LOAD_INIT_DATA';
 export const ADD_COMPONENT: string = 'ADD_COMPONENT';
-export const ADD_CHILD: string = 'ADD_CHILD';
-export const DELETE_CHILD: string = 'DELETE_CHILD';
 export const UPDATE_COMPONENT: string = 'UPDATE_COMPONENT';
 export const DELETE_COMPONENT: string = 'DELETE_COMPONENT';
+export const UPDATE_CHILDREN: string = 'UPDATE_CHILDREN';
+export const REASSIGN_PARENT: string = 'REASSIGN_PARENT';
+export const SELECT_PARENT: string = 'SELECT_PARENTS';
+export const OPEN_EXPANSION_PANEL: string = 'OPEN_EXPANSION_PANEL';
+export const MOVE_TO_TOP: string = 'MOVE_TO_TOP';
+export const MOVE_TO_BOTTOM: string = 'MOVE_TO_BOTTOM';
+export const DELETE_PROP: string = 'DELETE_PROP';
+export const ADD_PROP: string = 'ADD_PROP';
 export const CHANGE_FOCUS_COMPONENT: string = 'CHANGE_FOCUS_COMPONENT';
 export const CHANGE_COMPONENT_FOCUS_CHILD: string = 'CHANGE_COMPONENT_FOCUS_CHILD';
 export const CHANGE_FOCUS_CHILD: string = 'CHANGE_FOCUS_CHILD';
@@ -19,4 +25,4 @@ export const CREATE_APPLICATION_ERROR: string = 'CREATE_APPLICATION_ERROR';
 export const DELETE_ALL_DATA: string = 'DELETE_ALL_DATA';
 export const CHANGE_IMAGE_PATH: string = 'CHANGE_IMAGE_PATH';
 export const UPDATE_HTML_ATTR: string = 'UPDATE_HTML_ATTR';
-export const UPDATE_CHILDREN_SORT: string = 'UPDATE_CHILDREN_SORT'
+export const UPDATE_CHILDREN_SORT: string = 'UPDATE_CHILDREN_SORT';
