+++ tsconfig.json
@@ -3,14 +3,17 @@
     "outDir": "./build/",
     "sourceMap": true,
     "noImplicitAny": true,
+    "module": "commonjs",
+    "target": "es6",
     "jsx": "react",
-    "lib": ["dom", "esnext"],
-    "module": "esnext",
+    "lib": ["es6", "dom", "es2017"],
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true
   },
   "include": ["./src/**/*"],
   "exclude": ["node_modules"]
 }
+
+
